[neural-networks] renamed module for reductor
