+++ src/pymor/reductors/neural_network.py
